Replaced the typedefs in types.hpp with C++11 standard fixed width integer types (with the occassional warnings / correctness fix).

+++ xr3/h/XR/Rect.hpp
@@ -17,14 +17,14 @@ class Rect
 {
 public:
   // data
-  int16 x, y, w, h;
+  int32_t x, y, w, h;
 
   // structors
   Rect()
   : x(0), y(0), w(0), h(0)
   {}
 
-  Rect(int16 x_, int16 y_, int16 w_, int16 h_)
+  Rect(int32_t x_, int32_t y_, int32_t w_, int32_t h_)
   : x(x_), y(y_), w(w_), h(h_)
   {}
 };
